requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
jinja2==3.1.6
sqlalchemy==2.0.48
python-multipart==0.0.22
httpx==0.28.1
python-docx==1.2.0
alipay-sdk==2.0.0
wechatpayv3==2.0.2
itsdangerous==2.2.0
pytest==9.0.2
